## Description

This will update the OpenSearch Java client version from `v2.26.0` to
`v3.4.0`.

This upgrade requires addressing breaking API changes, updates to
functionality (such as including eager ordinals in the OS ISM response,
#26085), and updating tests and integration tests accordingly.

Because these modules are used as shared dependencies, the full upgrade
must be completed in a single step rather than split between
`zeebe/operate/tasklist` and `optimize` as previously planned.

## Checklist

- [ ] Ensure all tests pass

## Related issues

Closes to #40597, #26085, #40602

+++ optimize/backend/src/main/java/io/camunda/optimize/service/db/os/writer/DashboardWriterOS.java
@@ -140,7 +140,7 @@ public class DashboardWriterOS implements DashboardWriter {
 
     final UpdateResponse<Void> updateResponse = osClient.update(request, errorMessage);
 
-    if (updateResponse.shards().failed().intValue() > 0) {
+    if (updateResponse.shards().failed() > 0) {
       LOG.error(
           "Was not able to update dashboard with id [{}] and name [{}].", id, dashboard.getName());
       throw new OptimizeRuntimeException("Was not able to update dashboard!");
